requirements.txt
streamlit
Pillow
requests
python-dotenv
langchainhub
langchain_google_genai
docx
langchain